passes current UI tests.  more to come later

+++ src/main/java/makers/CookieMaker.java
@@ -68,7 +68,7 @@ public abstract class CookieMaker {
         return cps;
     }
 
-    public void setNum(long num) {
+    public void setNum(int num) {
         this.num = num;
     }
